+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - model's now expose a `log_likelihood_` property.
  - new `conditional_after` argument on `predict_*` methods that make prediction on censored subjects easier.
  - new `lifelines.utils.safe_exp` to make `exp` overflows easier to handle.
+ - smarter initial conditions for parametric regression models.
 
 ##### API changes
  - removed `lifelines.utils.gamma` - use `autograd_gamma` library instead.